+++ src/resources/manifest.php
@@ -2,5 +2,5 @@
 
 return [
     'name'    => 'Konekt AppShell Box',
-    'version' => '1.7.0'
+    'version' => '1.8.0'
 ];