+++ Changelog.md
@@ -1,5 +1,10 @@
 # AppShell Changelog
 
+## 1.8.0
+##### 2020-09-13
+
+- Added Laravel 8 Support
+
 ## 1.7.0
 ##### 2020-05-24